+++ weasyprint/css/__init__.py
@@ -26,13 +26,11 @@ from ..logger import LOGGER, PROGRESS_LOGGER
 from ..urls import URLFetchingError, get_url_attribute, url_join
 from . import counters, media_queries
 from .computed_values import COMPUTER_FUNCTIONS
+from .functions import check_var, parse_function
 from .properties import INHERITED, INITIAL_NOT_COMPUTED, INITIAL_VALUES, ZERO_PIXELS
+from .tokens import InvalidValues, Pending, get_url, remove_whitespace
 from .validation import preprocess_declarations
 from .validation.descriptors import preprocess_descriptors
-
-from .utils import (  # isort:skip
-    InvalidValues, Pending, check_var_function, get_url, parse_function,
-    remove_whitespace)
 
 # Reject anything not in here:
 PSEUDO_ELEMENTS = (
@@ -575,7 +573,7 @@ def declaration_precedence(origin, importance):
 
 def resolve_var(computed, token, parent_style, known_variables=None):
     """Return token with resolved CSS variables."""
-    if not check_var_function(token):
+    if not check_var(token):
         return
 
     if known_variables is None:
